+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+### Unreleased
+
+##### Added
+
+* Add selected list input to handle many to many associations.
+
 ### v0.11.0
 
 ##### Added